# fix(isPlainObject): Fix isPlainObject and perf 3x+