--- a/src/predicate/isPlainObject.spec.ts
+++ b/src/predicate/isPlainObject.spec.ts
@@ -1,113 +1,100 @@
-import { describe, expect, it } from 'vitest';
-import { isPlainObject } from './isPlainObject';
+import { describe, expect, it } from "vitest";
+import { isPlainObject } from "./isPlainObject";
+import { runInNewContext } from "node:vm";
 
-describe('isPlainObject', () => {
-  const str = 'key';
+describe("isPlainObject", () => {
+  it("should return true for plain objects", () => {
+    const str = "key";
 
-  function fnWithProto(x: number) {
-    // @ts-expect-error for the sake of testing
-    this.x = x;
-  }
+    expect(isPlainObject({})).toBe(true);
+    expect(isPlainObject(Object.create(null))).toBe(true);
+    expect(isPlainObject(new Object({ key: "new_object" }))).toBe(true);
+    expect(isPlainObject(new Object({ key: new Date() }))).toBe(true);
+    expect(isPlainObject({ 1: "integer_key" })).toBe(true);
+    expect(isPlainObject({ name: "string_key" })).toBe(true);
+    expect(isPlainObject({ [str]: "dynamic_string_key" })).toBe(true);
+    expect(isPlainObject({ [Symbol("tag")]: "value" })).toBe(true);
+    expect(
+      isPlainObject({
+        children: [{ key: "deep-children" }],
+        name: "deep-plain",
+      }),
+    ).toBe(true);
+    expect(
+      isPlainObject({
+        children: [{ key: new Date() }],
+        name: "deep-with-regular-object",
+      }),
+    ).toBe(true);
+    expect(isPlainObject({ constructor: { name: "Object2" } })).toBe(true);
+    expect(isPlainObject(JSON.parse("{}"))).toBe(true);
+    expect(isPlainObject(new Proxy({}, {}))).toBe(true);
+    expect(isPlainObject(new Proxy({ key: "proxied_key" }, {}))).toBe(true);
+  });
 
-  function ObjectConstructor() {}
+  it("should return false for invalid plain objects", () => {
+    function fnWithProto(x: number) {
+      // @ts-expect-error for the sake of testing
+      this.x = x;
+    }
 
-  ObjectConstructor.prototype.constructor = Object;
+    function ObjectConstructor() {}
 
-  const validPlainObjects = [
-    [{}, true],
-    [Object.create(null), true],
-    [new Object({ key: 'new_object' }), true],
-    [new Object({ key: new Date() }), true],
-    [{ 1: 'integer_key' }, true],
-    [{ name: 'string_key' }, true],
-    [{ [str]: 'dynamic_string_key' }, true],
-    [{ [Symbol('tag')]: 'value' }, true],
-    [{ children: [{ key: 'deep-children' }], name: 'deep-plain' }, true],
-    [{ children: [{ key: new Date() }], name: 'deep-with-regular-object' }, true],
-    [{ constructor: { name: 'Object2' } }, true],
-    [JSON.parse('{}'), true],
-    [new Proxy({}, {}), true],
-    [new Proxy({ key: 'proxied_key' }, {}), true],
-  ] as const;
+    ObjectConstructor.prototype.constructor = Object;
 
-  const invalidPlainObjects = [
-    ['hello', false],
-    [false, false],
-    [undefined, false],
-    [null, false],
-    [10, false],
-    [[], false],
-    [Number.NaN, false],
-    // functions and objects
-    [() => 'cool', false],
-    [new (class Cls {})(), false],
-    [new Intl.Locale('en'), false],
-    [new (class extends Object {})(), false],
-    [fnWithProto, false],
-    // Symbols
-    [Symbol('cool'), false],
-    [
-      {
-        [Symbol.iterator]: function* () {
-          yield 1;
-        },
+    expect(isPlainObject("hello")).toBe(false);
+    expect(isPlainObject(false)).toBe(false);
+    expect(isPlainObject(undefined)).toBe(false);
+    expect(isPlainObject(null)).toBe(false);
+    expect(isPlainObject(10)).toBe(false);
+    expect(isPlainObject([])).toBe(false);
+    expect(isPlainObject(Number.NaN)).toBe(false);
+    expect(isPlainObject(() => "cool")).toBe(false);
+    expect(isPlainObject(new (class Cls {})())).toBe(false);
+    expect(isPlainObject(new Intl.Locale("en"))).toBe(false);
+    expect(isPlainObject(new (class extends Object {})())).toBe(false);
+    expect(isPlainObject(fnWithProto)).toBe(false);
+    expect(isPlainObject(Symbol("cool"))).toBe(false);
+    expect(isPlainObject({
+      [Symbol.iterator]: function* () {
+        yield 1;
       },
+    })).toBe(false);
+    expect(isPlainObject({ [Symbol.toStringTag]: "string-tagged" })).toBe(
       false,
-    ],
-    [
-      {
-        [Symbol.toStringTag]: 'string-tagged',
-      },
-      false,
-    ],
-    // globalThis
-    [globalThis, false],
-    // Static built-in classes
-    // https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/JSON
-    [JSON, false],
-    [Math, false],
-    [Atomics, false],
-    [JSON, false],
-    // Built-in classes
-    [new Date(), false],
-    [new Map(), false],
-    [new Error(), false],
-    [new Set(), false],
-    [new Request('http://localhost'), false],
-    [new Promise(() => {}), false],
-    [Promise.resolve({}), false],
-    [Buffer.from('ABC'), false],
-    [new Uint8Array([1, 2, 3]), false],
-    [Object.create({}), false],
-    [/(\d+)/, false],
-    [new RegExp('/d+/'), false],
-    [/d+/, false],
-    // Template literals
-    [`cool`, false],
-    [String.raw`rawtemplate`, false],
-    // @ts-expect-error to allow testing crafted object
-    [new ObjectConstructor(), false],
-    [
+    );
+    expect(isPlainObject(globalThis)).toBe(false);
+    expect(isPlainObject(JSON)).toBe(false);
+    expect(isPlainObject(Math)).toBe(false);
+    expect(isPlainObject(Atomics)).toBe(false);
+    expect(isPlainObject(new Date())).toBe(false);
+    expect(isPlainObject(new Map())).toBe(false);
+    expect(isPlainObject(new Error())).toBe(false);
+    expect(isPlainObject(new Set())).toBe(false);
+    expect(isPlainObject(new Request("http://localhost"))).toBe(false);
+    expect(isPlainObject(new Promise(() => {}))).toBe(false);
+    expect(isPlainObject(Promise.resolve({}))).toBe(false);
+    expect(isPlainObject(Buffer.from("ABC"))).toBe(false);
+    expect(isPlainObject(new Uint8Array([1, 2, 3]))).toBe(false);
+    expect(isPlainObject(Object.create({}))).toBe(false);
+    expect(isPlainObject(/(\d+)/)).toBe(false);
+    expect(isPlainObject(new RegExp("/d+/"))).toBe(false);
+    expect(isPlainObject(/d+/)).toBe(false);
+    expect(isPlainObject(`cool`)).toBe(false);
+    expect(isPlainObject(String.raw`rawtemplate`)).toBe(false);
+    // eslint-disable-next-line
+    // @ts-ignore
+    expect(isPlainObject(new ObjectConstructor())).toBe(false);
+    expect(isPlainObject(
       new Proxy(new Date(), {
         get(target) {
           return target;
         },
       }),
-      false,
-    ],
-  ] as const;
-
-  const cases = [...validPlainObjects, ...invalidPlainObjects] as const;
-
-  it.each(cases)('when "%s" is given, should return %s', (v, expected) => {
-    expect(isPlainObject(v)).toStrictEqual(expected);
+    )).toBe(false);
   });
 
-  describe('Cross-realms node:vm:runInNewContext support', () => {
-    const isNodeLike = !('window' in globalThis);
-    it.skipIf(!isNodeLike)('should support vm', async () => {
-      const runInNewContext = await import('node:vm').then(mod => mod.runInNewContext);
-      expect(isPlainObject(runInNewContext('({})'))).toBe(true);
-    });
+  it("should return true for cross-realm plain objects", async () => {
+    expect(isPlainObject(runInNewContext("({})"))).toBe(true);
   });
 });