Hello, I'm really impressed by estoolkit. Great !

I'm porting few more tests and fixes to the isPlainObject implementation. Mostly taken from my own isPlainObject implementation: see https://github.com/belgattitude/httpx/tree/main/packages/plain-object#readme

This fix few bugs that I found. So I've completed the test suite. 

## Fixes

- #694 
- #693 
- #695 


## Performance

### "Realworld" scenario

Should also bring a >3x speedup based on my "real-world" scenarios

![image](https://github.com/user-attachments/assets/7537ea21-f9fe-4d03-8853-4c0d7f401ab5)

https://github.com/belgattitude/httpx/tree/main/packages/plain-object#benchmarks

### estoolkit bench

After the changes
```
 ✓ benchmarks/performance/isPlainObject.bench.ts > isPlainObject 3920ms
     name                                       hz     min     max    mean     p75     p99    p995    p999     rme  samples
   · es-toolkit/isPlainObject         2,519,342.19  0.0004  0.0325  0.0004  0.0004  0.0007  0.0009  0.0017  ±0.06%  1259672
   · es-toolkit/compat/isPlainObject  3,528,357.44  0.0003  0.2412  0.0003  0.0003  0.0005  0.0007  0.0014  ±0.12%  1764179   fastest
   · lodash/isPlainObject               692,780.40  0.0014  0.3287  0.0014  0.0014  0.0021  0.0026  0.0044
```

Before
```
✓ benchmarks/performance/isPlainObject.bench.ts > isPlainObject 3842ms
     name                                       hz     min     max    mean     p75     p99    p995    p999     rme  samples
   · es-toolkit/isPlainObject         2,496,468.28  0.0004  0.0388  0.0004  0.0004  0.0008  0.0010  0.0018  ±0.07%  1248235
   · es-toolkit/compat/isPlainObject  3,137,159.94  0.0003  0.3087  0.0003  0.0003  0.0008  0.0010  0.0019  ±0.20%  1568580   fastest
   · lodash/isPlainObject               702,054.71  0.0013  0.0356  0.0014  0.0014  0.0019  0.0025  0.0044  ±0.09%   351028   slowest
```

## Tasks

- [x] Improve test suite in https://github.com/toss/es-toolkit/pull/692/commits/ab76ab3b71b023e41dcb5a41b63492e1a8350641
- [x] Fix bugs: iterator symbol, anonymous classes, runInNewContext (see fixes below)
- [x] Refactor isPlainObject function

## Notes

> See the updated test suite in: https://github.com/toss/es-toolkit/pull/692/commits/ab76ab3b71b023e41dcb5a41b63492e1a8350641

- [x] If an object has an iterator symbol, it shouldn't be a plain-object (open to discuss ?)

![image](https://github.com/user-attachments/assets/648c22ac-982e-499a-a2eb-50d9019d7bc5)

- [x]  When passing an anonymous class it should not fail with TypeError 

![image](https://github.com/user-attachments/assets/f569e8b7-b026-4102-90a1-3c24492afb23)

- [x] Should support cross realms (node:vm runInNewContext)

![image](https://github.com/user-attachments/assets/063dd85e-9d20-4caa-b71b-d761b1f5d710)