@belgattitude 2024-10-08
Tests passes on my laptop... node 20.18.0, seems the CI doesn't

![image](https://github.com/user-attachments/assets/370db036-928f-4094-ab37-2506b1e90809)