@belgattitude 2024-10-08
This PR should make it compatible (drop in replacement) with

- https://github.com/sindresorhus/is-plain-obj
- https://github.com/belgattitude/httpx/tree/main/packages/plain-object#readme