@raon0211 2024-10-09
Hey, thanks for your contribution! I really appreciate your implementation—it’s much more modern and concise.

I made a few minor edits:

- I noticed that we check for `Symbol.toStringTag` to handle static objects like `JSON`, `Math`, and `Atomics`. I think we should consider them plain objects since they’re not instances of any class; they just aggregate some functions.
- Similarly, when checking for `Symbol.iterator` to handle `arguments`, I believe it should also be seen as a plain object. It is not an array, and mainly has plain properties like `0`, `1`, and `Symbol.iterator`.

Lastly, while these additional checks add complexity, they also increase the bundle size and may impact performance, even though they address some edge cases.