@belgattitude 2024-10-09
> I noticed that we check for Symbol.toStringTag to handle static objects like JSON, Math, and Atomics. I think we should consider them plain objects since they’re not instances of any class; they just aggregate some functions.

Thx @raon0211, I've tested the change on my repo. Not sure about static objects, imho they shouldn't be considered as plain objects.

Without the checks for static object, we get rid off 30 bytes (before - 101b -> 75b brotli/webpack)... The performance won't change a lot as the checks are done last. (I've benchmarked on my repo, and the impact isn't measurable). 

I'm happy with both ways, but I'd recommend to keep support for Atomics in general for "correctness" (if it actually exist :))