@belgattitude 2024-10-09
> Similarly, when checking for Symbol.iterator to handle arguments, I believe it should also be seen as a plain object. It is not an array, and mainly has plain properties like 0, 1, and Symbol.iterator.

That one I tend to agree (without this test the bundle would go from 101b -> 83b: in the scenario we keep toStringTag for atomics)