@raon0211 2024-10-09
Our team has been discussing the definition of `isPlainObject`. It seems the common understanding of plain objects in JavaScript includes the following points:

- They behave like typical object literals, such as `{ name: 'John' }`.
- Their prototype is either `Object.prototype` or `null`.
- When stringified using `toString`, they return `[object Object]`.

In light of this, we've updated our [documentation on plain objects](https://es-toolkit.slash.page/reference/predicate/isPlainObject.html) and revised the implementation accordingly. You can check out the changes [here](https://github.com/toss/es-toolkit/commit/755d9db62c3435ce6f2b85cfbf92ca96bde572e5).

We chose not to check for `Symbol.toStringTag` or `Symbol.iterator` for the following reasons:

- Not all internal objects have a `Symbol.toStringTag`. For instance, `Symbol.toStringTag in []` evaluates to false, even though `[].toString() === '[object Array]'`.
- There are valid plain objects that can include properties like `{ [Symbol.iterator]: function *() {} }`.